chore: rename source/ dir to src/
