+++ .changeset/honest-experts-notice.md
@@ -1,0 +1,5 @@
+---
+'react-medium-image-zoom': patch
+---
+
+rename source/ directory to src/
