+++ README.md
@@ -260,7 +260,7 @@ determining when to zoom and unzoom the component.
 
 You can import the default styles from `react-medium-image-zoom/dist/styles.css`
 and override the values from your code, or you can copy [the styles.css
-file](./source/styles.css) and alter it to your liking. The latter is the best
+file](./src/styles.css) and alter it to your liking. The latter is the best
 option, given `rem`s should be used instead of `px` to account for different
 default browser font sizes, and it's hard for a library to guess at what these
 values should be.
