+++ src/globals.d.ts
