+++ src/utils/adjust-svg-ids.test.ts
