+++ src/utils/adjust-svg-ids.ts
