+++ src/utils/get-div-img-style.test.ts
