+++ src/utils/get-div-img-style.ts
