+++ src/utils/get-img-alt.test.ts
