+++ src/utils/get-img-regular-style.test.ts
