+++ src/utils/get-img-regular-style.ts
