+++ src/utils/get-scale.ts
