+++ src/utils/get-style-ghost.test.ts
