+++ src/utils/get-style-ghost.ts
