+++ src/utils/get-style-modal-img.test.ts
