+++ src/utils/get-style-modal-img.ts
