+++ vitest.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'vitest/config'
 export default defineConfig({
   test: {
     environment: 'happy-dom',
-    include: ['source/**/*.test.{ts,tsx}'],
+    include: ['src/**/*.test.{ts,tsx}'],
     restoreMocks: true,
     setupFiles: ['./test/setup.ts'],
   },
